+++ tests/test_atableau.py
@@ -13,7 +13,7 @@
 # ]
 # ///
 
-## Requires: uv -- this installs the required python dependencies
+# Requires: uv -- this installs the required python dependencies
 
 r'''
 usage: test_atableau.py [-h] [-q] [-t THRESHOLD] [-w WORKERS] [-e | -i | -u] [files ...]
@@ -31,56 +31,54 @@ options:
 '''
 
 HELP = r'''
-This python script is part of the aTableau LaTeX package. It can be used
-to check that the output from the aTableau example files has not changed
-during development. The example files are extracted from the aTableau
-manual, atableau.tex, which is assumed to either be in the same
-directory as this script, or in the parent directory. If it does not
-already exist, the script creates a symbolic link that is the python
-equivalent of
+This python script is part of the aTableau LaTeX package. It can be used to
+check that the output from the aTableau example files has not changed during
+development. The example files are extracted from the aTableau manual,
+atableau.tex, which is assumed to either be in the same directory as this
+script, or in the parent directory. If it does not already exist, the script
+creates a symbolic link that is the python equivalent of
 
     ln -s ../atableau.tex atableau-examples.tex
 
-Running pdflatex on the file atableau-examples.tex extracts all of the
-examples in the manual, creating a separate file in this directory for
-each example. Running this script on the example files compares their
-output with the expected output.
+Running pdflatex on the file atableau-examples.tex extracts all of the examples
+in the manual, creating a separate file in this directory for each example.
+Running this script on the example files compares their output with the
+expected output.
 
 BEFORE starting development, the examples files should be initialised using
 
     test_examples.py -i
 
-This will extract the examples from the manual, and then create webp
-files for each example, such as ribbon-good.webp. These "good" webp
-files are used as the expected output of the examples.
+This will extract the examples from the manual, and then create webp files for
+each example, such as ribbon-good.webp. These "good" webp files are used as the
+expected output of the examples.
 
 Once the good files have been initialised, the command
 
     test_examples.py [files]
 
-compiles and tests all of the matching files to check changes. Here,
-<file> is interpreted liberally with wild-card expansions on both sides.
-For example,
+compiles and tests all of the matching files to check for changes. Here, <file>
+is interpreted liberally with wild-card expansions on both sides.  For example,
 
     test_examples.py tableau
 
-tests all of the example files with names that contain 'tableau'. When
-new examples are added to the manual, they can be extracted using:
+tests all of the example files with names that contain 'tableau'. When new
+examples are added to the manual, they can be extracted using:
 
     test_examples.py -e
 
-When they do ot already exist, this will also create a good webp for
-future comparisons, but it will not overwrite any existing good webp
-files for the examples.
+When they do ot already exist, this will also create a good webp files for
+future comparisons, but it will not overwrite any existing good webp files for
+the examples. (In fact, `-e` rewrites all of the example LaTeX files.)
 
-If any of the examples in the manual change in a good way, in the sense
-that the example if corrected or improved, then the good images can be
-updated using:
+If any of the examples in the manual change in a good way, in the sense that
+the example is corrected, or improved, then the good images can be updated
+using:
 
     test_examples -u [files]
 
-There are a large number of examples in the manual, but this is
-reasonably quick because they are processed in parallel.
+There are over 200 examples in the manual, but the script is reasonably quick
+because the examples are processed in parallel.
 
 Andrew Mathas
 October 2025
@@ -132,7 +130,6 @@ def red_text(text):
     '''
     return f'\033[31m{text}\033[39m'
 
-
 def example_number(file):
     '''
     Look in the example file `file`.tex to find the example number
@@ -143,7 +140,6 @@ def example_number(file):
                 break
 
     return line[1:].strip()
-
 
 def make_image(file, ext):
     '''
@@ -162,7 +158,6 @@ def make_image(file, ext):
     webp[0].save(f'{file}{ext}', 'WEBP')
     os.remove(f'{file}.pdf')
 
-
 def different_images(file, options):
     '''
     Return `True` or `False` depending on whether the image has changed
@@ -170,7 +165,6 @@ def different_images(file, options):
     diff = ImageChops.difference(Image.open(f'{file}.webp'), Image.open(f'{file}-good.webp'))
     diff_array = numpy.array(diff)
     mean_diff = diff_array.mean()
-
 
 def find_example_files(files):
     '''
@@ -191,7 +185,6 @@ def find_example_files(files):
             pass
 
     return example_files
-
 
 # ------------------------------------------------------------------------
 # actions
@@ -217,12 +210,10 @@ def extracting_image(file, options):
     if not os.path.isfile(f'{file}-good.webp'):
         initialising_image(file, options)
 
-
 def updating_image(file, options):
     '''
     Update the good webp image for file
     '''
-
     make_image(file, '.webp')
     if different_images(file, options):
         os.replace(f'{file}.webp', f'{file}-good.webp')
@@ -230,7 +221,6 @@ def updating_image(file, options):
             print(f' - {example_number(file):<14} updated ({file}.webp)')
     else:
         os.remove(f'{file}.webp')
-
 
 def checking_image(file, options):
     '''
@@ -243,7 +233,6 @@ def checking_image(file, options):
     elif not options.quiet:
         print(f' - {example_number(file):<14} OK ({file})')
         os.remove(f'{file}.webp')
-
 
 # ------------------------------------------------------------------------
 if __name__ == '__main__':